+++ dbms/derby/src/main/java/is/codion/dbms/derby/DerbyDatabaseFactory.java
@@ -40,7 +40,7 @@ public final class DerbyDatabaseFactory implements DatabaseFactory {
 	}
 
 	@Override
-	public Database createDatabase(String url) {
+	public Database create(String url) {
 		return new DerbyDatabase(url);
 	}