+++ dbms/hsqldb/src/main/java/is/codion/dbms/hsqldb/HSQLDatabaseFactory.java
@@ -36,7 +36,7 @@ public final class HSQLDatabaseFactory implements DatabaseFactory {
 	}
 
 	@Override
-	public Database createDatabase(String url) {
+	public Database create(String url) {
 		return new HSQLDatabase(url, Database.SELECT_FOR_UPDATE_NOWAIT.get());
 	}
 }