+++ dbms/oracle/src/main/java/is/codion/dbms/oracle/OracleDatabaseFactory.java
@@ -36,7 +36,7 @@ public final class OracleDatabaseFactory implements DatabaseFactory {
 	}
 
 	@Override
-	public Database createDatabase(String url) {
+	public Database create(String url) {
 		return new OracleDatabase(url, Database.SELECT_FOR_UPDATE_NOWAIT.get());
 	}
 }