+++ dbms/postgresql/src/main/java/is/codion/dbms/postgresql/PostgreSQLDatabaseFactory.java
@@ -36,7 +36,7 @@ public final class PostgreSQLDatabaseFactory implements DatabaseFactory {
 	}
 
 	@Override
-	public Database createDatabase(String url) {
+	public Database create(String url) {
 		return new PostgreSQLDatabase(url, Database.SELECT_FOR_UPDATE_NOWAIT.get());
 	}
 }