+++ dbms/sqlite/src/main/java/is/codion/dbms/sqlite/SQLiteDatabaseFactory.java
@@ -36,7 +36,7 @@ public final class SQLiteDatabaseFactory implements DatabaseFactory {
 	}
 
 	@Override
-	public Database createDatabase(String url) {
+	public Database create(String url) {
 		return new SQLiteDatabase(url);
 	}
 }